requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
allure-pytest==2.15.3
    # via playwright-tests (pyproject.toml)
allure-python-commons==2.15.3
    # via allure-pytest
annotated-types==0.7.0
    # via pydantic
attrs==25.4.0
    # via allure-python-commons
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
greenlet==3.3.1
    # via playwright
idna==3.11
    # via requests
iniconfig==2.3.0
    # via pytest
packaging==26.0
    # via pytest
playwright==1.58.0
    # via
    #   playwright-tests (pyproject.toml)
    #   pytest-playwright
pluggy==1.6.0
    # via
    #   allure-python-commons
    #   pytest
pydantic==2.12.5
    # via
    #   playwright-tests (pyproject.toml)
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via playwright-tests (pyproject.toml)
pyee==13.0.0
    # via playwright
pygments==2.19.2
    # via pytest
pyspellchecker==0.8.4
    # via playwright-tests (pyproject.toml)
pytest==9.0.2
    # via
    #   playwright-tests (pyproject.toml)
    #   allure-pytest
    #   pytest-base-url
    #   pytest-playwright
    #   pytest-repeat
pytest-base-url==2.1.0
    # via pytest-playwright
pytest-playwright==0.7.2
    # via playwright-tests (pyproject.toml)
pytest-repeat==0.9.4
    # via playwright-tests (pyproject.toml)
python-dotenv==1.2.1
    # via pydantic-settings
python-slugify==8.0.4
    # via pytest-playwright
requests==2.32.5
    # via pytest-base-url
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   pyee
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
